Master -> release-desktop-palau v1.16.0

+++ packages/sync-server/app/report/ReportRunner.js
@@ -1,10 +1,16 @@
 import config from 'config';
+import fs from 'fs';
+import path from 'path';
+import * as AWS from '@aws-sdk/client-s3';
+import mkdirp from 'mkdirp';
+
 import { COMMUNICATION_STATUSES } from 'shared/constants';
 import { getReportModule } from 'shared/reports';
 import { log } from 'shared/services/logging';
 import { createTupaiaApiClient, translateReportDataToSurveyResponses } from 'shared/utils';
 
-import { removeFile, createZippedExcelFile } from '../utils/files';
+import { removeFile, createZippedSpreadsheet, writeToSpreadsheet } from '../utils/files';
+import { getLocalisation } from '../localisation';
 
 export class ReportRunner {
   constructor(reportName, parameters, recipients, store, emailService) {
@@ -16,12 +22,14 @@ export class ReportRunner {
     this.tupaiaApiClient = null;
   }
 
-  validate(reportModule, reportDataGenerator) {
-    if (!config.mailgun.from) {
+  async validate(reportModule, reportDataGenerator) {
+    const localisation = await getLocalisation();
+
+    if (this.recipients.email && !config.mailgun.from) {
       throw new Error('ReportRunner - Email config missing');
     }
 
-    const disabledReports = config.localisation.data.disabledReports;
+    const { disabledReports } = localisation;
     if (disabledReports.includes(this.reportName)) {
       throw new Error(`ReportRunner - Report "${this.reportName}" is disabled`);
     }
@@ -37,7 +45,7 @@ export class ReportRunner {
     const reportModule = getReportModule(this.reportName);
     const reportDataGenerator = reportModule?.dataGenerator;
 
-    this.validate(reportModule, reportDataGenerator);
+    await this.validate(reportModule, reportDataGenerator);
 
     let reportData = null;
     try {
@@ -78,9 +86,47 @@ export class ReportRunner {
       await this.sendReportToTupaia(reportData);
       sent = true;
     }
+    if (this.recipients.s3) {
+      await this.sendReportToS3(reportData);
+      sent = true;
+    }
+    if (this.recipients.local) {
+      await this.sendReportToLocal(reportData);
+      sent = true;
+    }
     if (!sent) {
       throw new Error('ReportRunner - No recipients');
     }
+  }
+
+  /**
+   * @param reportData
+   * @returns {Promise<void>}
+   */
+  async sendReportToLocal(reportData) {
+    const reportName = this.getReportName();
+    for (const recipient of this.recipients.local) {
+      const { format, path: reportFolder } = recipient;
+      if (!format || !reportFolder) {
+        const str = JSON.stringify(recipient);
+        throw new Error(
+          `ReportRunner - local recipients must specifiy a format and a path, got: ${str}`,
+        );
+      }
+      await mkdirp(reportFolder);
+      const reportNameExtended = `${reportName}.${format}`;
+      const reportPath = path.resolve(reportFolder, reportNameExtended);
+      const outputPath = await writeToSpreadsheet(reportData, reportPath, format);
+      // eslint-disable-next-line no-console
+      console.log(outputPath);
+    }
+  }
+
+  /**
+   * @returns {string}
+   */
+  getReportName() {
+    return `${this.reportName}-report-${new Date().getTime()}`;
   }
 
   /**
@@ -90,11 +136,11 @@ export class ReportRunner {
    * @returns {Promise<void>}
    */
   async sendReportToEmail(reportData) {
-    const reportName = `${this.reportName}-report-${new Date().getTime()}`;
+    const reportName = this.getReportName();
 
-    let zipFile = null;
+    let zipFile;
     try {
-      zipFile = await createZippedExcelFile(reportName, reportData);
+      zipFile = await createZippedSpreadsheet(reportName, reportData);
 
       log.info(
         `ReportRunner - Sending report "${zipFile}" to "${this.recipients.email.join(',')}"`,
@@ -113,7 +159,7 @@ export class ReportRunner {
         throw new Error(`ReportRunner - Mailgun error: ${result.error}`);
       }
     } finally {
-      await removeFile(zipFile);
+      if (zipFile) await removeFile(zipFile);
     }
   }
 
@@ -138,5 +184,46 @@ export class ReportRunner {
     }
 
     await this.tupaiaApiClient.meditrak.createSurveyResponses(translated);
+  }
+
+  /**
+   * @param request ReportRequest
+   * @param reportData []
+   * @returns {Promise<void>}
+   */
+  async sendReportToS3(reportData) {
+    const { region, bucketName, bucketPath } = config.s3;
+
+    if (!bucketPath) {
+      throw new Error(`bucketPath must be set, e.g. 'au'`);
+    }
+
+    let zipFile;
+    const bookType = 'csv';
+    try {
+      zipFile = await createZippedSpreadsheet(this.reportName, reportData, bookType);
+
+      const filename = path.basename(zipFile);
+
+      log.info(
+        `ReportRunner - Uploading report to s3 "${bucketName}/${bucketPath}/${filename}" (${region})`,
+      );
+
+      const client = new AWS.S3({ region });
+
+      const fileStream = fs.createReadStream(zipFile);
+
+      await client.send(
+        new AWS.PutObjectCommand({
+          Bucket: bucketName,
+          Key: `${bucketPath}/${filename}`,
+          Body: fileStream,
+        }),
+      );
+
+      log.info(`ReportRunner - Uploaded report "${zipFile}" to s3`);
+    } finally {
+      if (zipFile) await removeFile(zipFile);
+    }
   }
 }
